eslint-config: promote future rules to main configs
Move the rules previously in @comet/eslint-config/future/* into the main react/nextjs/nestjs configs and remove the future subpath. The new rules now apply by default for the v9 release.
