+++ demo/admin/eslint.config.mjs
@@ -1,5 +1,5 @@
 import { defineConfig, globalIgnores } from "eslint/config";
-import eslintConfigReact from "@comet/eslint-config/future/react.js";
+import eslintConfigReact from "@comet/eslint-config/react.js";
 
 export default defineConfig([
     globalIgnores([
