+++ packages/admin/admin/eslint.config.mjs
@@ -1,5 +1,5 @@
 import { defineConfig, globalIgnores } from "eslint/config";
-import eslintConfigReact, { restrictedImportPaths } from "@comet/eslint-config/future/react.js";
+import eslintConfigReact, { restrictedImportPaths } from "@comet/eslint-config/react.js";
 // For more info, see https://github.com/storybookjs/eslint-plugin-storybook#configuration-flat-config-format
 import storybook from "eslint-plugin-storybook";
