+++ packages/eslint-config/nestjs.js
@@ -1,5 +1,12 @@
-import coreConfig from "./core.js";
+import coreConfig, { restrictedImportPatterns } from "./core.js";
 import globals from "globals";
+
+export const restrictedImportPaths = [
+    {
+        name: "node-cache",
+        message: "node-cache is abandonware. Use @nestjs/cache-manager instead",
+    },
+];
 
 /** @type {import('eslint')} */
 const config = [
@@ -22,6 +29,13 @@ const config = [
             "no-console": "off",
             "no-duplicate-imports": "error",
             "sort-imports": "off",
+            "no-restricted-imports": [
+                "error",
+                {
+                    paths: restrictedImportPaths,
+                    patterns: restrictedImportPatterns,
+                },
+            ],
         },
     },
 ];
