+++ packages/eslint-config/package.json
@@ -13,15 +13,6 @@
         "./core.js": {
             "import": "./core.js"
         },
-        "./future/nestjs.js": {
-            "import": "./future/nestjs.js"
-        },
-        "./future/nextjs.js": {
-            "import": "./future/nextjs.js"
-        },
-        "./future/react.js": {
-            "import": "./future/react.js"
-        },
         "./nestjs.js": {
             "import": "./nestjs.js"
         },
